feat: Only generate utils if they don't exist (#288)
* feat: Only generate utils when they don't exists

* run fix:prettier

---------

+++ plugins/typescript/src/generators/generateFetchers.ts
@@ -225,7 +225,10 @@ export const generateFetchers = async (context: Context, config: Config) => {
     }
   );
 
-  if (usedImportsKeys.includes("utils")) {
+  if (
+    !context.existsFile(`${utilsFilename}.ts`) &&
+    usedImportsKeys.includes("utils")
+  ) {
     await context.writeFile(`${utilsFilename}.ts`, getUtils());
   }
